Rename list_api to task_api
Most of the task submodules could potentially
contain the same classes so it is better that
they are all represented by one api module.
