+++ app/build.gradle
@@ -46,7 +46,7 @@ dependencies {
     implementation project(":core")
     implementation project(":features:stopwatch")
     implementation project(":features:task:list")
-    implementation project(":features:task:list_api")
+    implementation project(':features:task:task_api')
 
     implementation 'androidx.core:core-ktx:1.3.2'
     implementation 'androidx.appcompat:appcompat:1.2.0'
