+++ features/stopwatch/src/main/java/com/akjaw/stopwatch/presentation/StopwatchViewModel.kt
@@ -2,8 +2,8 @@ package com.akjaw.stopwatch.presentation
 
 import androidx.lifecycle.ViewModel
 import com.akjaw.stopwatch.domain.StopwatchListOrchestrator
-import com.akjaw.task.list_api.data.TaskRepository
-import com.akjaw.task.list_api.domain.Task
+import com.akjaw.task.api.data.TaskRepository
+import com.akjaw.task.api.domain.Task
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.MutableStateFlow
