+++ features/stopwatch/src/main/java/com/akjaw/stopwatch/view/AddStopwatchDialog.kt
@@ -16,8 +16,8 @@ import com.akjaw.core.common.view.PaddedDialog
 import com.akjaw.core.common.view.theme.TimiComposeTheme
 import com.akjaw.core.common.view.theme.taskShape
 import com.akjaw.core.common.view.theme.taskTextColorFor
-import com.akjaw.task.list_api.domain.Task
-import com.akjaw.task.list_api.view.tasksPreview
+import com.akjaw.task.api.domain.Task
+import com.akjaw.task.api.view.tasksPreview
 
 @Composable
 internal fun AddStopwatchDialog(
