+++ features/stopwatch/src/test/java/com/akjaw/stopwatch/domain/StopwatchListOrchestratorTest.kt
@@ -3,7 +3,7 @@ package com.akjaw.stopwatch.domain
 import androidx.compose.ui.graphics.Color
 import com.akjaw.core.common.domain.model.toTimestampMilliseconds
 import com.akjaw.stopwatch.domain.model.StopwatchState
-import com.akjaw.task.list_api.domain.Task
+import com.akjaw.task.api.domain.Task
 import io.mockk.every
 import io.mockk.mockk
 import kotlinx.coroutines.CoroutineScope
