+++ features/task/list/build.gradle
@@ -43,7 +43,7 @@ android {
 
 dependencies {
     implementation project(":core")
-    implementation project(":features:task:list_api")
+    implementation project(':features:task:task_api')
 
     implementation 'androidx.core:core-ktx:1.3.2'
     implementation 'androidx.appcompat:appcompat:1.2.0'
