+++ features/task/list/src/main/java/com/akjaw/task/list/composition/TaskListModule.kt
@@ -1,9 +1,9 @@
 package com.akjaw.task.list.composition
 
+import com.akjaw.task.api.data.TaskRepository
+import com.akjaw.task.api.view.TaskListScreenCreator
 import com.akjaw.task.list.data.TaskRepositoryImpl
 import com.akjaw.task.list.view.TaskListScreenCreatorImpl
-import com.akjaw.task.list_api.data.TaskRepository
-import com.akjaw.task.list_api.view.TaskListScreenCreator
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
