+++ features/task/list/src/main/java/com/akjaw/task/list/view/AddTaskFloatingActionButton.kt
@@ -29,7 +29,7 @@ import androidx.compose.ui.window.Dialog
 import com.akjaw.core.common.view.theme.TimiComposeTheme
 import com.akjaw.core.common.view.theme.taskColors
 import com.akjaw.core.common.view.theme.taskTextColorFor
-import com.akjaw.task.list_api.domain.Task
+import com.akjaw.task.api.domain.Task
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.launch
