--- features/task/list_api/src/main/java/com/akjaw/task/list_api/data/TaskRepository.kt
@@ -1,8 +1,0 @@
-package com.akjaw.task.list_api.data
-
-import com.akjaw.task.list_api.domain.Task
-
-interface TaskRepository {
-
-    val tasks: List<Task>
-}
