+++ features/task/task_api/src/main/AndroidManifest.xml
@@ -1,4 +1,4 @@
 <?xml version="1.0" encoding="utf-8"?>
-<manifest package="com.akjaw.task.list_api">
+<manifest package="com.akjaw.task.api">
 
 </manifest>
