+++ features/task/task_api/src/main/java/com/akjaw/task/api/data/TaskRepository.kt
@@ -1,0 +1,8 @@
+package com.akjaw.task.api.data
+
+import com.akjaw.task.api.domain.Task
+
+interface TaskRepository {
+
+    val tasks: List<Task>
+}
