+++ features/task/task_api/src/main/java/com/akjaw/task/api/domain/Task.kt
@@ -1,7 +1,7 @@
-package com.akjaw.task.list_api.domain
+package com.akjaw.task.api.domain
 
 import androidx.compose.ui.graphics.Color
-//TODO move to tasks:list-api
+
 data class Task(
     val name: String,
     val backgroundColor: Color, //TODO change the type?
