+++ features/task/task_api/src/main/java/com/akjaw/task/api/view/TaskViewUtils.kt
@@ -1,7 +1,7 @@
-package com.akjaw.task.list_api.view
+package com.akjaw.task.api.view
 
 import androidx.compose.ui.graphics.Color
-import com.akjaw.task.list_api.domain.Task
+import com.akjaw.task.api.domain.Task
 
 val tasksPreview = listOf(
     Task("Task 1", Color(132, 212, 240), false),
