+++ settings.gradle
@@ -13,4 +13,4 @@ include ':features:stopwatch'
 include ':core'
 include ':features:task'
 include ':features:task:list'
-include ':features:task:list_api'
+include ':features:task:task_api'
